requirements.txt
Django>=3.2,<4.0
psycopg2>=2.8
djangorestframework
djangorestframework-simplejwt
django-filter
requests
beautifulsoup4
pandas
xlsxwriter
drf-yasg